divided the code into modules
